requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
yt-dlp>=2024.1.0
